Moved all mustache classes into one package. No need for multiple packages inside a small module.
