+++ modules/lang-mustache/src/main/java/org/elasticsearch/script/mustache/MultiSearchTemplateAction.java
@@ -17,28 +17,28 @@
  * under the License.
  */
 
-package org.elasticsearch.action.search.template;
+package org.elasticsearch.script.mustache;
 
 import org.elasticsearch.action.Action;
-import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.client.ElasticsearchClient;
 
-public class SearchTemplateAction extends Action<SearchTemplateRequest, SearchTemplateResponse, SearchTemplateRequestBuilder> {
+public class MultiSearchTemplateAction
+        extends Action<MultiSearchTemplateRequest, MultiSearchTemplateResponse, MultiSearchTemplateRequestBuilder> {
 
-    public static final SearchTemplateAction INSTANCE = new SearchTemplateAction();
-    public static final String NAME = "indices:data/read/search/template";
+    public static final MultiSearchTemplateAction INSTANCE = new MultiSearchTemplateAction();
+    public static final String NAME = "indices:data/read/msearch/template";
 
-    private SearchTemplateAction() {
+    private MultiSearchTemplateAction() {
         super(NAME);
     }
 
     @Override
-    public SearchTemplateRequestBuilder newRequestBuilder(ElasticsearchClient client) {
-        return new SearchTemplateRequestBuilder(client, this);
+    public MultiSearchTemplateResponse newResponse() {
+        return new MultiSearchTemplateResponse();
     }
 
     @Override
-    public SearchTemplateResponse newResponse() {
-        return new SearchTemplateResponse();
+    public MultiSearchTemplateRequestBuilder newRequestBuilder(ElasticsearchClient client) {
+        return new MultiSearchTemplateRequestBuilder(client, this);
     }
 }
