+++ modules/lang-mustache/src/test/resources/org/elasticsearch/script/mustache/simple-msearch-template.json
